Parsing requirements line to fix an issue . Perfectly working setup.py parsing
