+++ pkg/parser/custom/utils/regex_utils/regex.go
@@ -1,0 +1,17 @@
+package regex_utils
+
+import (
+	"regexp"
+	"sync"
+)
+
+var cache sync.Map
+
+func MustCompileAndCache(exp string) *regexp.Regexp {
+	compiled, ok := cache.Load(exp)
+	if !ok {
+		compiled, _ = cache.LoadOrStore(exp, regexp.MustCompile(exp))
+	}
+
+	return compiled.(*regexp.Regexp)
+}
